Moving old examples to benchmarks.
Major migration from old examples to new format. Removing old examples to new directory, where they will be repurposed into benchmarks.
