+++ benchmarks/benchmark_obstacle_avoid_problem.py
