+++ benchmarks/configs/planar_quadrotor_system.yaml
