+++ benchmarks/configs/point_mass_system.yaml
