+++ benchmarks/ingredients/backward_reach_ingredient.py
@@ -4,11 +4,11 @@ import gym
 
 from gym_socks.envs.dynamical_system import DynamicalSystem
 
-from gym_socks.envs.sample import sample
-from gym_socks.envs.sample import sample_generator
-from gym_socks.envs.sample import step_sampler
-from gym_socks.envs.sample import uniform_grid
-from gym_socks.envs.sample import uniform_grid_step_sampler
+from gym_socks.sampling import sample
+from gym_socks.sampling import sample_generator
+from gym_socks.sampling import step_sampler
+from gym_socks.sampling import uniform_grid
+from gym_socks.sampling import uniform_grid_step_sampler
 
 import numpy as np
