+++ benchmarks/ingredients/matplotlibrc
