+++ docs/benchmarks/benchmark_obstacle_avoid_problem/index.rst
