+++ docs/benchmarks/example_separating_kernel/index.rst
