+++ docs/benchmarks/ingredients/plotting_ingredient/index.rst
